requirements.txt
antlr4-python3-runtime==4.13.1
PyQt6==6.6.1
